+++ src/redturtle/chefcookie/browser/viewlets.py
@@ -3,9 +3,15 @@ from plone import api
 from plone.app.layout.viewlets.common import ViewletBase
 from plone.memoize import ram
 from redturtle.chefcookie.interfaces import IChefCookieSettings
+from redturtle.chefcookie.defaults import domain_allowed
 from time import time
-
 import pkg_resources
+import six
+
+if six.PY2:
+    from urlparse import urlparse
+else:
+    from urllib.parse import urlparse
 
 
 CHEFCOOKIE_URL = "{portal_url}/++resource++redturtle.chefcookie/chefcookie/chefcookie.min.js?v={version}"
@@ -24,8 +30,14 @@ class GetChefcookieJs(ViewletBase):
 
     def have_chefcookie_configuration(self):
         try:
-            self.context.portal_registry.forInterface(IChefCookieSettings)
-            return True
+            cc = self.context.portal_registry.forInterface(IChefCookieSettings)
+
+            if cc.enable_cc and domain_allowed(
+                cc.domain_whitelist, urlparse(self.request.get("URL")).netloc
+            ):
+                return True
+
+            return False
         except Exception:
             return False
 
@@ -52,7 +64,8 @@ class GetChefcookieJs(ViewletBase):
 
     def get_css_link(self):
         return "{portal_url}/++resource++redturtle.chefcookie/styles.css?v={version}".format(
-            portal_url=api.portal.get().portal_url(), version=self.get_version()
+            portal_url=api.portal.get().portal_url(),
+            version=self.get_version(),
         )
 
     def get_type(self):