+++ src/redturtle/chefcookie/upgrades.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
 
 import logging
-
+from plone import api
+from redturtle.chefcookie.interfaces import IChefCookieSettings
 
 logger = logging.getLogger(__name__)
 
@@ -15,3 +16,8 @@ def update_profile(context, profile, run_dependencies=True):
 def update_registry(context):
     update_profile(context, "plone.app.registry", run_dependencies=False)
     logger.info("Update registry")
+
+
+def update_to_1002(context):
+    update_registry(context)
+    api.portal.set_registry_record("enable_cc", True, interface=IChefCookieSettings)